Analyzer/requirements.txt
aniso8601==10.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
deep-translator==1.11.4
filelock==3.17.0
Flask==3.1.0
Flask-RESTful==0.3.10
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
moondream==0.0.5
mpmath==1.3.0
numpy==2.2.3
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pillow==10.4.0
protobuf==6.30.0
pyreadline3==3.5.4
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sympy==1.13.3
tokenizers==0.20.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
